Don't hardcode the path to the GitNub application in the nub CLI tool.
I didn't want to install GitNub in /Applications, which caused the nub
command to fail terribly.  This change uses the standard cocoa
NSWorkspace API for opening applications with paths.

Note that this changes causes a slightly different, but more
``normal'' behavior for applications.

If it is desirable to have multiple repos open at the same time, that
should be handled in the application itself.  This invocation via nub
should already do the right thing to accommodate that.